requirements.txt
# Taleemabad Observability Dashboard
# For Railway deployment

# Core Framework
streamlit>=1.32.0

# Data Visualization
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0

# Database Connections
psycopg2-binary>=2.9.9

# Google Cloud (BigQuery)
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0

# Utilities
python-dotenv>=1.0.0
